requirements.txt
djangorestframework
markdown
django-filter
pika
pydantic
dependency-injector
requests
psycopg2